Moved css files to styles folder
