+++ e-commerce-app/src/CartProduct.js
@@ -1,9 +1,9 @@
 import { useContext } from "react";
 import CartContext from "./CartContext";
-
-import "./common.css"
-import "./CartProduct.css";
 import { Link } from "react-router-dom";
+
+import "./Styles/common.css"
+import "./Styles/CartProduct.css";
 
 const CartProduct = (props) => {
   const { cartUpdater } = useContext(CartContext);
