+++ e-commerce-app/src/ProductCard.js
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 
-import "./ProductCard.css";
+import "./Styles/ProductCard.css";
 
 const ProductCard = (props) => {
   return (
