+++ e-commerce-app/src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+
+import './Styles/index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
